Grown box for the MOL hydro fluxes was incorrect wrt F90 version
This fix ensures that the MOL hydro fluxes are computed on the same
extents that were used in the F90 version of the code. The replaced
code was computing on a box one cell short in the transverse
directions. This shortens the box by one only in the direction
currently considered for the fluxes.