Fix mismatch between EB tiling and no tiling.
Before this fix, if one ran the code with and without tiling and EB,
one would get different results (fcompare would show diffs of order 1e-4
after 1 time step of the EB MMS test case). This was present in the
F90 and CPP versions of the code: F90 without tiling = CPP without
tiling and F90 with tiling = CPP with tiling BUT F90/CPP with tiling !=
F90/CPP with tiling.

This was easy to test in the CPP version of the code as one only had
to try `fabarray.mfiter_tile_size = 1024 16 16` or
`fabarray.mfiter_tile_size = 1024 1024 1024`. For the F90 version, one
has to change `hydro_tile_size`.

This problem was also evident in the fact that with tiling on the EB
MMS test case was not _perfectly_ symmetric as it should have been.

The reason for this fix is that all the grow cells should be filled
when using EB (it is expected by the algorithm).