Update the EB redistribution (#274)
This PR updates the EB redistribution to use IAMR redistribution
utils. The big deal with this PR is that we get to use State
Redistribution, as described by Berger and Giuliani (2021). This is a
better way to perform EB redistribution. I performed extensive
verification of the new method and believe that it works as expected
and performs better than existing redistribution. Specifically:

- we can safely set eb_small_vfrac to zero for all cases
- EB cases perform just as well as existing redistribution, some better
- EB-C7 and EB-C8 no longer need a specific eb_small_vfrac (using 0 works)
- EB-C14 (shock interacting with ramp) runs (does not run with existing)
- EB-C13 is a new case that is added

New options are documented in the code.